backend/requirements.txt
Flask==2.3.3
chromadb==0.4.18
PyPDF2==3.0.1
pandas==2.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
